Add ClustergrammerCard to file details page
